# aura-api/openapi.yaml
openapi: 3.0.3
info:
  title: Aura API
  version: "1.0.0"
  description: >
    API Aura — Mobile ↔ API ↔ Agent (Miroir).
    Auth côté utilisateur via **JWT Bearer**.
    Auth côté agent via **ApiKey** + **x-device-id**.
servers:
  - url: http://127.0.0.1:3000/api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Authorization: Bearer <JWT>'
    apiKeyAgent:
      type: apiKey
      in: header
      name: Authorization
      description: 'Authorization: ApiKey <CLE_DU_DEVICE>'
    xDeviceId:
      type: apiKey
      in: header
      name: x-device-id
      description: 'UUID du device (agent)'
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
      example: { error: Unauthorized, message: Invalid credentials }

    # ── Health & Public
...
    User:
      type: object
      required: [id, email]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        firstName: { type: string, nullable: true }
        lastName: { type: string, nullable: true }
      example:
        id: 37bfce36-e569-46dd-b6b3-7f3c1b98c668
        email: alice@example.com
        firstName: Alice
        lastName: Martin
...
    DeviceListItem:
      type: object
      required: [id, name, createdAt, disabled]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        createdAt: { type: string, format: date-time }
        disabled: { type: boolean }
        online: { type: boolean, nullable: true }
        lastSeenAt: { type: string, format: date-time, nullable: true }
      example:
        id: 89e81262-2101-4f6a-9969-40b81a18d929
        name: Miroir Salon
        createdAt: 2025-08-24T15:37:27.390Z
        disabled: false
        online: true
        lastSeenAt: 2025-08-27T18:37:00.000Z
    DeviceSimple:
      type: object
      required: [id, name]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
      example: { id: 89e81262-2101-4f6a-9969-40b81a18d929, name: Miroir Salon }
    DevicePairBody:
      type: object
      required: [deviceId, pairingToken]
      properties:
        deviceId: { type: string, format: uuid }
        pairingToken: { type: string, minLength: 4 }
      example:
        deviceId: 89e81262-2101-4f6a-9969-40b81a18d929
        pairingToken: "123456"
    DeviceRenameBody:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
      example: { name: Miroir Chambre }

    # ── Pairing (Agent)
...
    LedState:
      type: object
      required: [on, color, brightness]
      properties:
        on: { type: boolean }
        color: { type: string, pattern: '^#[0-9A-Fa-f]{6}$' }
        brightness: { type: integer, minimum: 0, maximum: 100 }
        preset: { type: string, nullable: true }
      example: { on: true, color: "#00A3FF", brightness: 42, preset: ocean }
...
    MusicState:
      type: object
      required: [status, volume]
      properties:
        status: { type: string, enum: [play, pause] }
        volume: { type: integer, minimum: 0, maximum: 100 }
        track: { type: object, nullable: true }
      example: { status: pause, volume: 50, track: null }
...
    WidgetItem:
      type: object
      required: [key, enabled, orderIndex]
      properties:
        key: { type: string, enum: [clock, weather, music, leds] }
        enabled: { type: boolean }
        orderIndex: { type: integer, minimum: 0 }
        config: {}
      example: { key: weather, enabled: true, orderIndex: 1, config: { city: Paris, units: metric } }
    WidgetsList:
      type: array
      items: { $ref: '#/components/schemas/WidgetItem' }
      example:
        - { key: clock, enabled: true, orderIndex: 0, config: { format: "24h" } }
        - { key: weather, enabled: true, orderIndex: 1, config: { city: Paris, units: metric } }
        - { key: music, enabled: true, orderIndex: 2, config: {} }
    WidgetsPutBody:
      type: object
      required: [items]
      properties:
        items:
          type: array
          minItems: 1
          items: { $ref: '#/components/schemas/WidgetItem' }
      example:
        items:
          - { key: clock, enabled: true, orderIndex: 0, config: { format: "24h" } }
          - { key: weather, enabled: true, orderIndex: 1, config: { city: Paris, units: metric } }
    WidgetsPutResponse:
      type: object
      required: [items]
      properties:
        items: { $ref: '#/components/schemas/WidgetsList' }

    # ── Weather
...
    DeviceOwnerResponse:
      type: object
      required: [owner]
      properties:
        owner:
          oneOf:
            - $ref: '#/components/schemas/User'
            - type: "null"
      example:
        owner:
          id: 37bfce36-e569-46dd-b6b3-7f3c1b98c668
          email: alice@example.com
          firstName: Alice
          lastName: Martin
    DeviceOnlineResponse:
      type: object
      required: [online]
      properties:
        online: { type: boolean }
        lastSeenAt:
          type: string
          format: date-time
          nullable: true
      example: { online: true, lastSeenAt: 2025-08-27T18:37:00.000Z }
    UnpairResponse:
      type: object
      required: [ok]
      properties:
        ok: { type: boolean }
      example: { ok: true }
...
paths:
...
  /devices:
    get:
      tags: [Devices]
      summary: List my devices
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/DeviceListItem' }
  /devices/pair:
    post:
      tags: [Devices, Pairing]
      summary: Pair a device with a short token (supports transfer tokens)
      description: >
        Associe un device via un token court généré par l'agent. Si le token
        a été créé avec `transfer=true`, un transfert de propriétaire est autorisé.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DevicePairBody' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [device]
                properties:
                  device: { $ref: '#/components/schemas/DeviceSimple' }
        '400':
          description: no active token
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '401':
          description: invalid pairing token
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '409':
          description: already paired or disabled
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '410':
          description: token expired
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  /devices/{deviceId}:
    put:
      tags: [Devices]
      summary: Rename device
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/DeviceRenameBody' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [device]
                properties:
                  device: { $ref: '#/components/schemas/DeviceSimple' }
    delete:
      tags: [Devices]
      summary: Delete device (hard delete)
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '204':
          description: no content
  /devices/{deviceId}/state:
    get:
      tags: [Devices]
      summary: Device snapshot (leds, music, widgets)
      description: >
        Accessible via **JWT propriétaire** ou via **ApiKey + x-device-id** de l'agent.
      security:
        - bearerAuth: []
        - apiKeyAgent: []
          xDeviceId: []
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema:
                type: object
                required: [leds, music, widgets]
                properties:
                  leds: { $ref: '#/components/schemas/LedState' }
                  music: { $ref: '#/components/schemas/MusicState' }
                  widgets: { $ref: '#/components/schemas/WidgetsList' }
  /devices/{deviceId}/owner:
    get:
      tags: [Devices]
      summary: Get device owner (or null)
      description: >
        Accessible via **JWT propriétaire** ou via **ApiKey + x-device-id** de l'agent.
      security:
        - bearerAuth: []
        - apiKeyAgent: []
          xDeviceId: []
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceOwnerResponse' }
  /devices/{deviceId}/online:
    get:
      tags: [Devices]
      summary: Get device online status
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DeviceOnlineResponse' }
  /devices/{deviceId}/unpair:
    post:
      tags: [Devices, Pairing]
      summary: Unpair device (set ownerId=null)
      description: >
        Accessible via **JWT propriétaire** ou via **ApiKey + x-device-id** de l'agent.
      security:
        - bearerAuth: []
        - apiKeyAgent: []
          xDeviceId: []
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/UnpairResponse' }

  # NEW: register device without user (agentless bootstrap)
...
  /devices/{deviceId}/widgets:
    get:
      tags: [Widgets]
      summary: List device widgets
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WidgetsList' }
    put:
      tags: [Widgets]
      summary: Upsert device widgets
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: deviceId
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/WidgetsPutBody' }
      responses:
        '200':
          description: ok
          content:
            application/json:
              schema: { $ref: '#/components/schemas/WidgetsPutResponse' }

  # 7) Weather
...